+++ server/lib/job-queue/handlers/move-to-object-storage.ts
@@ -6,8 +6,8 @@ import { createTorrentAndSetInfoHash } from '@server/helpers/webtorrent'
 import { CONFIG } from '@server/initializers/config'
 import { HLS_STREAMING_PLAYLIST_DIRECTORY } from '@server/initializers/constants'
 import { storeHLSFile, storeWebTorrentFile } from '@server/lib/object-storage'
-import { moveToNextState } from '@server/lib/video'
 import { getHLSDirectory, getHlsResolutionPlaylistFilename } from '@server/lib/video-paths'
+import { moveToNextState } from '@server/lib/video-state'
 import { VideoModel } from '@server/models/video/video'
 import { VideoJobInfoModel } from '@server/models/video/video-job-info'
 import { MVideoFile, MVideoWithAllFiles } from '@server/types/models'
